o Fix repo group REST client and util
